feat: Add a grid version to Diary view
