+++ src/components/Home/Home.tsx
@@ -1,36 +1,16 @@
-import { useMDState, useMDDispatch } from "@/config";
-import type {
-  MediaDiaryState,
-  UserFuegoValidated,
-  MediaDiaryWithId,
-} from "@/types";
-import { createPosterURL } from "@/utils";
-import {
-  Box,
-  Button,
-  Flex,
-  Grid,
-  Heading,
-  HStack,
-  Square,
-  useColorMode,
-  Text,
-  Image,
-} from "@chakra-ui/react";
-import dayjs from "dayjs";
-import React, { useState, useEffect } from "react";
-import { useSWRInfinite, cache } from "swr";
-import { fuegoDiaryGet } from "./config";
-import type { SortType } from "./config";
-import { LogoIcon, StarEmptyIcon, FilmIcon, AlbumIcon, TvIcon } from "@/icons";
+import { useMDState } from "@/config";
+import { LogoIcon } from "@/icons";
 import { MdLoader } from "@/md";
-import { StarIcon, ArrowDownIcon } from "@chakra-ui/icons";
-import Rating from "react-rating";
+import type { MediaDiaryWithId, UserFuegoValidated } from "@/types";
+import { ArrowDownIcon } from "@chakra-ui/icons";
+import { Button, Heading, HStack, Square } from "@chakra-ui/react";
+import React, { useEffect, useState } from "react";
+import { cache, useSWRInfinite } from "swr";
 import { HomeHeader } from "./components";
-
-interface ListState {
-  [key: string]: MediaDiaryState;
-}
+import DiaryGrid from "./components/DiaryGrid";
+import DiaryList from "./components/DiaryList";
+import type { SortType } from "./config";
+import { fuegoDiaryGet } from "./config";
 
 export default function Home({
   user,
@@ -38,12 +18,11 @@ export default function Home({
   user: UserFuegoValidated;
 }): JSX.Element {
   const state = useMDState();
-  const dispatch = useMDDispatch();
-  const { colorMode } = useColorMode();
   const [sortType, setSortType] = useState<SortType>({
     type: "diaryDate",
     sort: "desc",
   });
+  const [viewType, setViewType] = useState<"grid" | "list">("list");
 
   const { data, error, size, setSize, mutate } = useSWRInfinite<
     MediaDiaryWithId[]
@@ -110,193 +89,17 @@ export default function Home({
 
   if (data) {
     const allData = data.flat();
-    const diaryDates: ListState = allData.reduce<ListState>((a, c) => {
-      if (c.diaryDate) {
-        const dateString = dayjs(c.diaryDate).format("YYYY-MM");
-        a[dateString] = Object.assign({ ...a[dateString] }, { [c.id]: c });
-      }
-      return a;
-    }, {});
-
     const isReachingEnd =
       isEmpty || (data && data[data.length - 1]?.length < 30);
     return (
       <>
-        <HomeHeader sortType={sortType} onChange={(val) => setSortType(val)} />
-        {Object.keys(diaryDates).map((month, monthIndex) => {
-          return (
-            <Grid
-              templateColumns={{
-                base: "3rem 1fr",
-                md: "0.1fr 0.9fr",
-              }}
-              key={monthIndex}
-              borderLeftWidth={{ base: 0, md: "1px" }}
-              borderRightWidth={{ base: 0, md: "1px" }}
-              px={{ md: 8 }}
-            >
-              <Box>
-                <Text
-                  fontSize={{ base: "lg", md: "2xl" }}
-                  color={colorMode === "light" ? "gray.600" : "gray.300"}
-                  fontWeight="bold"
-                  position="sticky"
-                  top="7.5rem"
-                >
-                  {dayjs(`${month}-01`).format("MMM")}
-                  <Text
-                    as="span"
-                    display="block"
-                    fontSize="sm"
-                    color="purple.700"
-                  >
-                    {dayjs(`${month}-01`).format("YYYY")}
-                  </Text>
-                </Text>
-              </Box>
-              <Box>
-                {Object.keys(diaryDates[month]).map((day, dayIndex) => {
-                  const {
-                    rating,
-                    title,
-                    poster,
-                    releasedDate,
-                    artist,
-                    type,
-                    season,
-                    seenEpisodes,
-                  } = diaryDates[month][day];
-                  const diaryDate = diaryDates[month][day].diaryDate;
-                  return (
-                    <Grid
-                      gridTemplateColumns={{
-                        base: "1.5rem 4rem 1fr",
-                        md: "3rem 7rem 1fr",
-                      }}
-                      gridGap="1rem"
-                      borderBottomWidth="1px"
-                      px={3}
-                      py={{ base: 4, md: 5 }}
-                      key={monthIndex + dayIndex}
-                      _hover={{
-                        bg: colorMode === "light" ? "purple.50" : "gray.700",
-                        cursor: "pointer",
-                      }}
-                      onClick={() =>
-                        dispatch({
-                          type: "day",
-                          payload: diaryDates[month][day],
-                        })
-                      }
-                    >
-                      <Box>
-                        <Text
-                          fontSize={{ base: "lg", md: "xl" }}
-                          color={
-                            colorMode === "light" ? "gray.500" : "gray.300"
-                          }
-                        >
-                          {diaryDate
-                            ? new Date(diaryDate).toLocaleDateString("en-us", {
-                                day: "numeric",
-                              })
-                            : "No Date"}
-                        </Text>
-                      </Box>
-                      <Box>
-                        <Image
-                          src={createPosterURL(poster, type)}
-                          ignoreFallback
-                          borderRadius="5px"
-                          border="1px solid"
-                          borderColor="gray.300"
-                        />
-                      </Box>
-                      <Flex flexDirection="column">
-                        <Text fontSize={{ base: "md", md: "xl" }}>{title}</Text>
-                        <Text fontSize={{ base: "sm" }} color="gray.500" pb={2}>
-                          {new Date(releasedDate).toLocaleDateString("en-US", {
-                            year: "numeric",
-                          })}
-                          <Text as="span" px={2}>
-                            ·
-                          </Text>
-                          {artist}
-                        </Text>
-                        {type === "tv" && (
-                          <Flex
-                            py={2}
-                            fontSize="sm"
-                            color="gray.500"
-                            fontStyle="italic"
-                          >
-                            <Text>
-                              <Text as="span" fontWeight="semibold">
-                                S:{" "}
-                              </Text>
-                              {season}
-                            </Text>
-                            {typeof seenEpisodes !== "undefined" &&
-                              seenEpisodes?.length > 0 && (
-                                <>
-                                  <Text px={2}>·</Text>
-                                  <Text>
-                                    <Text as="span" fontWeight="semibold">
-                                      Ep.{" "}
-                                    </Text>
-                                    {seenEpisodes
-                                      ?.sort((a, b) => (a < b ? -1 : 1))
-                                      .join(", ")}
-                                  </Text>
-                                </>
-                              )}
-                          </Flex>
-                        )}
-                        <Flex
-                          mt="auto"
-                          alignItems="center"
-                          justifyContent="space-between"
-                        >
-                          <Box mt={type === "tv" ? "-3px" : undefined}>
-                            {rating === 0 ? (
-                              <Text fontSize="sm" color="gray.500">
-                                No Rating
-                              </Text>
-                            ) : (
-                              <Rating
-                                fractions={2}
-                                readonly
-                                initialRating={rating}
-                                fullSymbol={
-                                  <StarIcon
-                                    h={{ base: "12px", md: "20px" }}
-                                    w={{ base: "12px", md: "20px" }}
-                                    color="purple.400"
-                                  />
-                                }
-                                emptySymbol={
-                                  // eslint-disable-next-line react/jsx-no-undef
-                                  <StarEmptyIcon
-                                    h={{ base: "12px", md: "20px" }}
-                                    w={{ base: "12px", md: "20px" }}
-                                    stroke="purple.400"
-                                  />
-                                }
-                              />
-                            )}
-                          </Box>
-                          {type === "movie" && <FilmIcon />}
-                          {type === "album" && <AlbumIcon />}
-                          {type === "tv" && <TvIcon />}
-                        </Flex>
-                      </Flex>
-                    </Grid>
-                  );
-                })}
-              </Box>
-            </Grid>
-          );
-        })}
+        <HomeHeader
+          sortType={sortType}
+          onChange={(val) => setSortType(val)}
+          view={{ type: viewType, onChange: (val) => setViewType(val) }}
+        />
+        {viewType === "list" && <DiaryList data={allData} />}
+        {viewType === "grid" && <DiaryGrid data={allData} />}
         {!isReachingEnd && (
           <HStack mt={3} mb={10} borderLeftWidth="1px" borderRightWidth="1px">
             <Button
