+++ src/components/Home/components/DiaryList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DiaryList";
