+++ src/components/Icons/ListIcon.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import type { IconProps } from "@chakra-ui/react";
+import { Icon } from "@chakra-ui/react";
+
+export default function ListIcon(props: IconProps): JSX.Element {
+  return (
+    <Icon
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      {...props}
+    >
+      <line x1="8" y1="6" x2="21" y2="6"></line>
+      <line x1="8" y1="12" x2="21" y2="12"></line>
+      <line x1="8" y1="18" x2="21" y2="18"></line>
+      <line x1="3" y1="6" x2="3.01" y2="6"></line>
+      <line x1="3" y1="12" x2="3.01" y2="12"></line>
+      <line x1="3" y1="18" x2="3.01" y2="18"></line>
+    </Icon>
+  );
+}
