+++ src/components/Icons/index.ts
@@ -13,3 +13,5 @@ export { default as StarEmptyIcon } from "./StarEmptyIcon";
 export { default as TvIcon } from "./TvIcon";
 export { default as TwitterIcon } from "./TwitterIcon";
 export { default as MoonAltIcon } from "./MoonAltIcon";
+export { default as ListIcon } from "./ListIcon";
+export { default as GridIcon } from "./GridIcon";
